Extract `ContextHistory`
